UcmHalOutStream.cpp: Async MMAP write finally works
The tinyalsa version currently in my jellybean tree is not compatible
with the "upstream" version from https://github.com/tinyalsa/tinyalsa.git
that I use in test binary build. This is due to added avail_min parameter
that needs to be initialized for tinyalsa to work correctly.

However, if the async buffer handling is implemented outside tinyalsa
there is no problem as long as pcm_mmap_write() call is not blocked
due to the DMA buffer getting full.

Signed-off-by: Jyri Sarha <jsarha@ti.com>